+++ packages/fury-adapter-oas3-parser/lib/context.js
@@ -3,11 +3,14 @@ const State = require('./state.js');
 class Context {
   constructor(namespace, options) {
     this.namespace = namespace;
+    this.options = options || {};
 
-    if (options === undefined) {
-      this.options = { generateSourceMap: false };
-    } else {
-      this.options = options;
+    if (this.options.generateSourceMap === undefined) {
+      this.options.generateSourceMap = false;
+    }
+
+    if (this.options.generateMessageBody === undefined) {
+      this.options.generateMessageBody = true;
     }
 
     this.state = new State();